fix: we cant register links async, so split things out to keep imports down in links.ts
