+++ src/Components/ServiceScene/Breakdowns/Patterns/PatternsLogsSampleScene.tsx
@@ -15,18 +15,17 @@ import React from 'react';
 import { LoadingState } from '@grafana/data';
 import { Alert, Button } from '@grafana/ui';
 import {
-  getFieldsVariable,
-  getLevelsVariable,
-  getLineFilterVariable,
   LOG_STREAM_SELECTOR_EXPR,
   PATTERNS_SAMPLE_SELECTOR_EXPR,
   VAR_PATTERNS_EXPR,
 } from '../../../../services/variables';
-import { buildDataQuery, LokiQuery, renderPatternFilters } from '../../../../services/query';
+import { buildDataQuery, renderPatternFilters } from '../../../../services/query';
 import { getQueryRunner } from '../../../../services/panel';
 import { AppliedPattern } from '../../../IndexScene/IndexScene';
 import { PatternsViewTableScene } from './PatternsViewTableScene';
 import { emptyStateStyles } from '../FieldsBreakdownScene';
+import { getFieldsVariable, getLevelsVariable, getLineFilterVariable } from '../../../../services/variableGetters';
+import { LokiQuery } from '../../../../services/lokiQuery';
 
 interface PatternsLogsSampleSceneState extends SceneObjectState {
   pattern: string;
