+++ src/module.tsx
@@ -1,17 +1,14 @@
 import { lazy } from 'react';
 import { AppPlugin } from '@grafana/data';
+import { linkConfigs } from 'services/extensions/links';
 
 const App = lazy(async () => {
   const { wasmSupported } = await import('services/sorting');
-  const { linkConfigs } = await import('services/extensions/links');
+
   const { default: initRuntimeDs } = await import('services/datasource');
   const { default: init } = await import('@bsull/augurs');
 
   initRuntimeDs();
-
-  for (const linkConfig of linkConfigs) {
-    plugin.configureExtensionLink(linkConfig);
-  }
 
   if (wasmSupported()) {
     await init();
@@ -21,3 +18,7 @@ const App = lazy(async () => {
 });
 
 export const plugin = new AppPlugin<{}>().setRootPage(App);
+
+for (const linkConfig of linkConfigs) {
+  plugin.configureExtensionLink(linkConfig);
+}
