+++ src/services/extensions/links.ts
@@ -1,10 +1,10 @@
 import { PluginExtensionLinkConfig, PluginExtensionPanelContext, PluginExtensionPoints } from '@grafana/data';
-import { LabelType } from 'services/fields';
-import { getMatcherFromQuery } from 'services/logql';
 
-import { LokiQuery } from 'services/query';
-import { appendUrlParameter, createAppUrl, setUrlParameter, UrlParameters } from 'services/routing';
-import { SERVICE_NAME } from 'services/variables';
+import { SERVICE_NAME, VAR_DATASOURCE, VAR_FIELDS, VAR_LABELS } from 'services/variables';
+import pluginJson from '../../plugin.json';
+import { LokiQuery } from '../lokiQuery';
+import { getMatcherFromQuery } from '../logqlMatchers';
+import { LabelType } from '../fieldsTypes';
 
 const title = 'Open in Explore Logs';
 const description = 'Open current query in the Explore Logs view';
@@ -68,4 +68,35 @@ function contextToLink<T extends PluginExtensionPanelContext>(context?: T) {
   return {
     path: createAppUrl(`/explore/service/${serviceName}/logs`, params),
   };
+}
+
+export function createAppUrl(path = '/explore', urlParams?: URLSearchParams): string {
+  return `/a/${pluginJson.id}${path}${urlParams ? `?${urlParams.toString()}` : ''}`;
+}
+
+export const UrlParameters = {
+  DatasourceId: `var-${VAR_DATASOURCE}`,
+  TimeRangeFrom: 'from',
+  TimeRangeTo: 'to',
+  Labels: `var-${VAR_LABELS}`,
+  Fields: `var-${VAR_FIELDS}`,
+} as const;
+export type UrlParameterType = (typeof UrlParameters)[keyof typeof UrlParameters];
+
+export function setUrlParameter(key: UrlParameterType, value: string, initalParams?: URLSearchParams): URLSearchParams {
+  const searchParams = new URLSearchParams(initalParams?.toString() ?? location.search);
+  searchParams.set(key, value);
+
+  return searchParams;
+}
+
+export function appendUrlParameter(
+  key: UrlParameterType,
+  value: string,
+  initalParams?: URLSearchParams
+): URLSearchParams {
+  const searchParams = new URLSearchParams(initalParams?.toString() ?? location.search);
+  searchParams.append(key, value);
+
+  return searchParams;
 }
