+++ src/services/fields.ts
@@ -10,8 +10,6 @@ import {
 import { getColorByIndex } from './scenes';
 import { AddToFiltersButton, VariableFilterType } from 'Components/ServiceScene/Breakdowns/AddToFiltersButton';
 import {
-  getLogsStreamSelector,
-  getValueFromFieldsFilter,
   LEVEL_VARIABLE_VALUE,
   LOG_STREAM_SELECTOR_EXPR,
   LogsQueryOptions,
@@ -25,6 +23,8 @@ import { map, Observable } from 'rxjs';
 import { SortBy, SortByScene } from '../Components/ServiceScene/Breakdowns/SortByScene';
 import { getDetectedFieldsFrame } from '../Components/ServiceScene/ServiceScene';
 import { averageFields } from '../Components/ServiceScene/Breakdowns/FieldsBreakdownScene';
+import { getLogsStreamSelector, getValueFromFieldsFilter } from './variableGetters';
+import { LabelType } from './fieldsTypes';
 
 export type DetectedLabel = {
   label: string;
@@ -167,13 +167,6 @@ export function selectFrameTransformation(frame: DataFrame) {
       })
     );
   };
-}
-
-// copied from public/app/plugins/datasource/loki/types.ts
-export enum LabelType {
-  Indexed = 'I',
-  StructuredMetadata = 'S',
-  Parsed = 'P',
 }
 
 export function getLabelTypeFromFrame(labelKey: string, frame: DataFrame, index = 0): null | LabelType {
