+++ src/services/fieldsTypes.ts
@@ -1,0 +1,6 @@
+// copied from public/app/plugins/datasource/loki/types.ts
+export enum LabelType {
+  Indexed = 'I',
+  StructuredMetadata = 'S',
+  Parsed = 'P',
+}
