+++ src/services/levels.ts
@@ -1,9 +1,10 @@
 import { DataFrame } from '@grafana/data';
 import { SeriesVisibilityChangeMode } from '@grafana/ui';
-import { getLevelsVariable, LEVEL_VARIABLE_VALUE } from './variables';
+import { LEVEL_VARIABLE_VALUE } from './variables';
 import { SceneObject } from '@grafana/scenes';
 import { FilterOp } from './filters';
 import { addToFilters, replaceFilter } from 'Components/ServiceScene/Breakdowns/AddToFiltersButton';
+import { getLevelsVariable } from './variableGetters';
 
 /**
  * Given a set of `visibleLevels` in a panel, it returns a list of the new visible levels
