+++ src/services/routing.ts
@@ -137,34 +137,3 @@ export function buildServicesRoute(extraQueryParams?: UrlQueryMap): UrlQueryMap 
     ...extraQueryParams,
   };
 }
-
-export function createAppUrl(path = '/explore', urlParams?: URLSearchParams): string {
-  return `/a/${pluginJson.id}${path}${urlParams ? `?${urlParams.toString()}` : ''}`;
-}
-
-export const UrlParameters = {
-  DatasourceId: `var-${VAR_DATASOURCE}`,
-  TimeRangeFrom: 'from',
-  TimeRangeTo: 'to',
-  Labels: `var-${VAR_LABELS}`,
-  Fields: `var-${VAR_FIELDS}`,
-} as const;
-export type UrlParameterType = (typeof UrlParameters)[keyof typeof UrlParameters];
-
-export function setUrlParameter(key: UrlParameterType, value: string, initalParams?: URLSearchParams): URLSearchParams {
-  const searchParams = new URLSearchParams(initalParams?.toString() ?? location.search);
-  searchParams.set(key, value);
-
-  return searchParams;
-}
-
-export function appendUrlParameter(
-  key: UrlParameterType,
-  value: string,
-  initalParams?: URLSearchParams
-): URLSearchParams {
-  const searchParams = new URLSearchParams(initalParams?.toString() ?? location.search);
-  searchParams.append(key, value);
-
-  return searchParams;
-}
