+++ src/services/variableGetters.ts
@@ -1,0 +1,171 @@
+import {
+  AdHocFiltersVariable,
+  CustomVariable,
+  DataSourceVariable,
+  sceneGraph,
+  SceneObject,
+  SceneVariableState,
+} from '@grafana/scenes';
+import { CustomConstantVariable } from './CustomConstantVariable';
+import {
+  AdHocFieldValue,
+  FieldValue,
+  JSON_FORMAT_EXPR,
+  LOGS_FORMAT_EXPR,
+  LogsQueryOptions,
+  MIXED_FORMAT_EXPR,
+  SERVICE_NAME,
+  VAR_DATASOURCE,
+  VAR_FIELD_GROUP_BY,
+  VAR_FIELDS,
+  VAR_FIELDS_EXPR,
+  VAR_LABEL_GROUP_BY,
+  VAR_LABELS,
+  VAR_LABELS_EXPR,
+  VAR_LEVELS,
+  VAR_LEVELS_EXPR,
+  VAR_LINE_FILTER,
+  VAR_LINE_FILTER_EXPR,
+  VAR_PATTERNS,
+  VAR_PATTERNS_EXPR,
+  VAR_SERVICE,
+} from './variables';
+import { AdHocVariableFilter } from '@grafana/data';
+
+export function getServiceSelectionStringVariable(sceneRef: SceneObject) {
+  const variable = sceneGraph.lookupVariable(VAR_SERVICE, sceneRef);
+  if (!(variable instanceof CustomConstantVariable)) {
+    throw new Error('VAR_SERVICE not found');
+  }
+  return variable;
+}
+
+export function getPatternsVariable(scene: SceneObject) {
+  const variable = sceneGraph.lookupVariable(VAR_PATTERNS, scene);
+  if (!(variable instanceof CustomVariable)) {
+    throw new Error('VAR_PATTERNS not found');
+  }
+  return variable;
+}
+
+export function getLabelsVariable(scene: SceneObject) {
+  return getAdHocFiltersVariable(VAR_LABELS, scene);
+}
+
+export function getFieldsVariable(scene: SceneObject) {
+  return getAdHocFiltersVariable(VAR_FIELDS, scene);
+}
+
+export function getLevelsVariable(scene: SceneObject) {
+  return getAdHocFiltersVariable(VAR_LEVELS, scene);
+}
+
+export function getLineFilterVariable(scene: SceneObject) {
+  const variable = sceneGraph.lookupVariable(VAR_LINE_FILTER, scene);
+  if (!(variable instanceof CustomVariable)) {
+    throw new Error('VAR_LINE_FILTER not found');
+  }
+  return variable;
+}
+
+export function getLabelGroupByVariable(scene: SceneObject) {
+  const variable = sceneGraph.lookupVariable(VAR_LABEL_GROUP_BY, scene);
+  if (!(variable instanceof CustomConstantVariable)) {
+    throw new Error('VAR_LABEL_GROUP_BY not found');
+  }
+  return variable;
+}
+
+export function getFieldGroupByVariable(scene: SceneObject) {
+  const variable = sceneGraph.lookupVariable(VAR_FIELD_GROUP_BY, scene);
+  if (!(variable instanceof CustomConstantVariable)) {
+    throw new Error('VAR_FIELD_GROUP_BY not found');
+  }
+  return variable;
+}
+
+export function getDataSourceVariable(scene: SceneObject) {
+  const variable = sceneGraph.lookupVariable(VAR_DATASOURCE, scene);
+  if (!(variable instanceof DataSourceVariable)) {
+    throw new Error('VAR_DATASOURCE not found');
+  }
+  return variable;
+}
+
+export function getAdHocFiltersVariable(variableName: string, scene: SceneObject) {
+  const variable = sceneGraph.lookupVariable(variableName, scene);
+
+  if (!(variable instanceof AdHocFiltersVariable)) {
+    throw new Error(`Could not get AdHocFiltersVariable ${variableName}. Variable not found.`);
+  }
+  return variable;
+}
+
+export function getLogsStreamSelector(options: LogsQueryOptions) {
+  const {
+    labelExpressionToAdd = '',
+    structuredMetadataToAdd = '',
+    fieldExpressionToAdd = '',
+    parser = undefined,
+  } = options;
+
+  switch (parser) {
+    case 'structuredMetadata':
+      return `{${VAR_LABELS_EXPR}${labelExpressionToAdd}} ${structuredMetadataToAdd} ${VAR_LEVELS_EXPR} ${VAR_PATTERNS_EXPR} ${VAR_LINE_FILTER_EXPR}`;
+    case 'json':
+      return `{${VAR_LABELS_EXPR}${labelExpressionToAdd}} ${structuredMetadataToAdd} ${VAR_LEVELS_EXPR} ${VAR_PATTERNS_EXPR} ${VAR_LINE_FILTER_EXPR} ${JSON_FORMAT_EXPR} ${fieldExpressionToAdd} ${VAR_FIELDS_EXPR}`;
+    case 'logfmt':
+      return `{${VAR_LABELS_EXPR}${labelExpressionToAdd}} ${structuredMetadataToAdd} ${VAR_LEVELS_EXPR} ${VAR_PATTERNS_EXPR} ${VAR_LINE_FILTER_EXPR} ${LOGS_FORMAT_EXPR} ${fieldExpressionToAdd} ${VAR_FIELDS_EXPR}`;
+    default:
+      return `{${VAR_LABELS_EXPR}${labelExpressionToAdd}} ${structuredMetadataToAdd} ${VAR_LEVELS_EXPR} ${VAR_PATTERNS_EXPR} ${VAR_LINE_FILTER_EXPR} ${MIXED_FORMAT_EXPR} ${fieldExpressionToAdd} ${VAR_FIELDS_EXPR}`;
+  }
+}
+
+export function getUrlParamNameForVariable(variableName: string) {
+  return `var-${variableName}`;
+}
+
+export function getValueFromFieldsFilter(filter: AdHocVariableFilter, variableName: string = VAR_FIELDS): FieldValue {
+  try {
+    return JSON.parse(filter.value);
+  } catch (e) {
+    console.error(`Failed to parse ${variableName}`, e);
+    throw e;
+  }
+}
+
+export function getValueFromAdHocVariableFilter(
+  variable: AdHocFiltersVariable,
+  filter?: AdHocVariableFilter
+): AdHocFieldValue {
+  if (variable.state.name === VAR_FIELDS && filter) {
+    return getValueFromFieldsFilter(filter);
+  }
+
+  return {
+    value: filter?.value,
+  };
+}
+
+export function getServiceName(scene: SceneObject) {
+  const labelsVariable = getLabelsVariable(scene);
+  return getServiceNameFromVariableState(labelsVariable.state);
+}
+
+export function getServiceNameFromVariableState(
+  adHocFiltersVariableState: SceneVariableState & { filters: AdHocVariableFilter[] }
+) {
+  const serviceName = adHocFiltersVariableState.filters
+    .filter((filter) => filter.key === SERVICE_NAME)
+    .map((filter) => filter.value);
+
+  if (!serviceName) {
+    throw new Error('Service present in filters selected');
+  }
+  return serviceName[0];
+}
+
+export function getDataSourceName(scene: SceneObject) {
+  const dsVariable = getDataSourceVariable(scene);
+  return dsVariable.getValue();
+}
