+++ src/services/variables.ts
@@ -1,14 +1,21 @@
-import {
-  AdHocFiltersVariable,
-  CustomVariable,
-  DataSourceVariable,
-  sceneGraph,
-  SceneObject,
-  SceneVariableState,
-} from '@grafana/scenes';
-import { CustomConstantVariable } from './CustomConstantVariable';
-import { AdHocVariableFilter } from '@grafana/data';
-import { AdHocFieldValue, FieldValue } from '../Components/ServiceScene/Breakdowns/AddToFiltersButton';
+export interface FieldValue {
+  value: string;
+  parser: ParserType;
+}
+
+export interface AdHocFieldValue {
+  value?: string;
+  parser?: ParserType;
+}
+
+export type ParserType = 'logfmt' | 'json' | 'mixed' | 'structuredMetadata';
+
+export type LogsQueryOptions = {
+  labelExpressionToAdd?: string;
+  structuredMetadataToAdd?: string;
+  fieldExpressionToAdd?: string;
+  parser?: ParserType;
+};
 
 export const VAR_LABELS = 'filters';
 export const VAR_LABELS_EXPR = '${filters}';
@@ -40,150 +47,3 @@ export const ALL_VARIABLE_VALUE = '$__all';
 export const LEVEL_VARIABLE_VALUE = 'detected_level';
 export const SERVICE_NAME = 'service_name';
 export const EMPTY_VARIABLE_VALUE = '""';
-
-export type ParserType = 'logfmt' | 'json' | 'mixed' | 'structuredMetadata';
-
-export type LogsQueryOptions = {
-  labelExpressionToAdd?: string;
-  structuredMetadataToAdd?: string;
-  fieldExpressionToAdd?: string;
-  parser?: ParserType;
-};
-
-export function getLogsStreamSelector(options: LogsQueryOptions) {
-  const {
-    labelExpressionToAdd = '',
-    structuredMetadataToAdd = '',
-    fieldExpressionToAdd = '',
-    parser = undefined,
-  } = options;
-
-  switch (parser) {
-    case 'structuredMetadata':
-      return `{${VAR_LABELS_EXPR}${labelExpressionToAdd}} ${structuredMetadataToAdd} ${VAR_LEVELS_EXPR} ${VAR_PATTERNS_EXPR} ${VAR_LINE_FILTER_EXPR}`;
-    case 'json':
-      return `{${VAR_LABELS_EXPR}${labelExpressionToAdd}} ${structuredMetadataToAdd} ${VAR_LEVELS_EXPR} ${VAR_PATTERNS_EXPR} ${VAR_LINE_FILTER_EXPR} ${JSON_FORMAT_EXPR} ${fieldExpressionToAdd} ${VAR_FIELDS_EXPR}`;
-    case 'logfmt':
-      return `{${VAR_LABELS_EXPR}${labelExpressionToAdd}} ${structuredMetadataToAdd} ${VAR_LEVELS_EXPR} ${VAR_PATTERNS_EXPR} ${VAR_LINE_FILTER_EXPR} ${LOGS_FORMAT_EXPR} ${fieldExpressionToAdd} ${VAR_FIELDS_EXPR}`;
-    default:
-      return `{${VAR_LABELS_EXPR}${labelExpressionToAdd}} ${structuredMetadataToAdd} ${VAR_LEVELS_EXPR} ${VAR_PATTERNS_EXPR} ${VAR_LINE_FILTER_EXPR} ${MIXED_FORMAT_EXPR} ${fieldExpressionToAdd} ${VAR_FIELDS_EXPR}`;
-  }
-}
-
-export function getPatternsVariable(scene: SceneObject) {
-  const variable = sceneGraph.lookupVariable(VAR_PATTERNS, scene);
-  if (!(variable instanceof CustomVariable)) {
-    throw new Error('VAR_PATTERNS not found');
-  }
-  return variable;
-}
-
-export function getLabelsVariable(scene: SceneObject) {
-  return getAdHocFiltersVariable(VAR_LABELS, scene);
-}
-
-export function getFieldsVariable(scene: SceneObject) {
-  return getAdHocFiltersVariable(VAR_FIELDS, scene);
-}
-
-export function getLevelsVariable(scene: SceneObject) {
-  return getAdHocFiltersVariable(VAR_LEVELS, scene);
-}
-
-export function getLineFilterVariable(scene: SceneObject) {
-  const variable = sceneGraph.lookupVariable(VAR_LINE_FILTER, scene);
-  if (!(variable instanceof CustomVariable)) {
-    throw new Error('VAR_LINE_FILTER not found');
-  }
-  return variable;
-}
-
-export function getLabelGroupByVariable(scene: SceneObject) {
-  const variable = sceneGraph.lookupVariable(VAR_LABEL_GROUP_BY, scene);
-  if (!(variable instanceof CustomConstantVariable)) {
-    throw new Error('VAR_LABEL_GROUP_BY not found');
-  }
-  return variable;
-}
-
-export function getFieldGroupByVariable(scene: SceneObject) {
-  const variable = sceneGraph.lookupVariable(VAR_FIELD_GROUP_BY, scene);
-  if (!(variable instanceof CustomConstantVariable)) {
-    throw new Error('VAR_FIELD_GROUP_BY not found');
-  }
-  return variable;
-}
-
-export function getDataSourceVariable(scene: SceneObject) {
-  const variable = sceneGraph.lookupVariable(VAR_DATASOURCE, scene);
-  if (!(variable instanceof DataSourceVariable)) {
-    throw new Error('VAR_DATASOURCE not found');
-  }
-  return variable;
-}
-
-export function getAdHocFiltersVariable(variableName: string, scene: SceneObject) {
-  const variable = sceneGraph.lookupVariable(variableName, scene);
-
-  if (!(variable instanceof AdHocFiltersVariable)) {
-    throw new Error(`Could not get AdHocFiltersVariable ${variableName}. Variable not found.`);
-  }
-  return variable;
-}
-
-export function getServiceSelectionStringVariable(sceneRef: SceneObject) {
-  const variable = sceneGraph.lookupVariable(VAR_SERVICE, sceneRef);
-  if (!(variable instanceof CustomConstantVariable)) {
-    throw new Error('VAR_SERVICE not found');
-  }
-  return variable;
-}
-
-export function getUrlParamNameForVariable(variableName: string) {
-  return `var-${variableName}`;
-}
-
-export function getValueFromFieldsFilter(filter: AdHocVariableFilter, variableName: string = VAR_FIELDS): FieldValue {
-  try {
-    return JSON.parse(filter.value);
-  } catch (e) {
-    console.error(`Failed to parse ${variableName}`, e);
-    throw e;
-  }
-}
-
-export function getValueFromAdHocVariableFilter(
-  variable: AdHocFiltersVariable,
-  filter?: AdHocVariableFilter
-): AdHocFieldValue {
-  if (variable.state.name === VAR_FIELDS && filter) {
-    return getValueFromFieldsFilter(filter);
-  }
-
-  return {
-    value: filter?.value,
-  };
-}
-
-export function getServiceName(scene: SceneObject) {
-  const labelsVariable = getLabelsVariable(scene);
-  return getServiceNameFromVariableState(labelsVariable.state);
-}
-
-export function getServiceNameFromVariableState(
-  adHocFiltersVariableState: SceneVariableState & { filters: AdHocVariableFilter[] }
-) {
-  const serviceName = adHocFiltersVariableState.filters
-    .filter((filter) => filter.key === SERVICE_NAME)
-    .map((filter) => filter.value);
-
-  if (!serviceName) {
-    throw new Error('Service present in filters selected');
-  }
-  return serviceName[0];
-}
-
-export function getDataSourceName(scene: SceneObject) {
-  const dsVariable = getDataSourceVariable(scene);
-  return dsVariable.getValue();
-}
