+++ tests/exploreServicesBreakDown.spec.ts
@@ -2,8 +2,8 @@ import { expect, test } from '@grafana/plugin-e2e';
 import { ExplorePage, PlaywrightRequest } from './fixtures/explore';
 import { testIds } from '../src/services/testIds';
 import { mockEmptyQueryApiResponse } from './mocks/mockEmptyQueryApiResponse';
-import { LokiQuery } from '../src/services/query';
 import exp = require('node:constants');
+import { LokiQuery } from '../src/services/lokiQuery';
 
 const fieldName = 'caller';
 const levelName = 'detected_level';
